Complete restructuring of test files
The Unit Test will now determine which functions are present in the
current libasm.a, and will only test these functions. It will also
inform the user which functions were not found inside of their libasm.a.

Mandatory and Bonus are now tested with the same command, and if bonus
functions are absent the user will simply be informed of their absence.

The non-criterion tests still require that all 6 mandatory functions are
present in libasm.a.